habilitando paginacao e cache

+++ forum-example/src/main/java/br/com/forumexample/controller/TopicosController.java
@@ -1,11 +1,17 @@
 package br.com.forumexample.controller;
 
 import java.net.URI;
-import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort.Direction;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -15,6 +21,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
 
@@ -40,16 +47,20 @@ public class TopicosController {
 	private CursoRepository cursoRepository;
 
 	@GetMapping /* recebendo parametro url */
-	public List<TopicoDto> lista(String nomeCurso) {
-
+	@Cacheable(value = "listaDeTopicos")
+	public Page<TopicoDto> lista(@RequestParam(required = false) String nomeCurso, 
+			@PageableDefault(sort = "id", direction = Direction.DESC, page = 0, size = 10) Pageable paginacao) {
+		
+		//Pageable paginacao = PageRequest.of(pagina, qtd, Direction.DESC, ordenacao);
+		
 		if (nomeCurso == null) {
-			List<Topico> topicos = topicoRepository
-					.findAll();/*
+			Page<Topico> topicos = topicoRepository
+					.findAll(paginacao);/*
 								 * metodo que faz a consulta carregando todos os registros do banco de dados
 								 */
 			return TopicoDto.converter(topicos);
 		} else { /* padrao de nomenclatura do springdata */
-			List<Topico> topicos = topicoRepository.findByCurso_Nome(nomeCurso);
+			Page<Topico> topicos = topicoRepository.findByCurso_Nome(nomeCurso, paginacao);
 			return TopicoDto.converter(topicos);
 		}
 
@@ -58,6 +69,7 @@ public class TopicosController {
 	/* para pegar o parametro no corpo da requisição */
 	@PostMapping
 	@Transactional
+	@CacheEvict(value = "listaDeTopicos", allEntries = true) /*limpar o cache*/
 	public ResponseEntity<TopicoDto> cadastrar(@RequestBody @Valid TopicoForm form, UriComponentsBuilder uriBuilder) {
 		Topico topico = form
 				.converter(cursoRepository);/* criar um objeto topico, pois estamos recebendo um topicoForm */
@@ -76,28 +88,45 @@ public class TopicosController {
 
 	/* lógica para detalhar um topico */
 	@GetMapping("/{id}") /* anotação para dizer que é uma variavel da url */
-	public DetalhesDoTopicoDto detalhar(@PathVariable Long id) {
-		Topico topico = topicoRepository.getOne(id);
-		return new DetalhesDoTopicoDto(topico);
+	public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id) {
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if (topico.isPresent()) {
+			return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));
+		}
+		
+		return ResponseEntity.notFound().build();
 	}
 
 	@PutMapping("/{id}") /* utilizado para sobrescrever o recurso */
 	@Transactional /*avisar que é pra commitar a atualização*/
+	@CacheEvict(value = "listaDeTopicos", allEntries = true) /*limpar o cache*/
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
 		
-		/*está vindo um ID do topico que quero atualizar, carregar ele do banco de dados
-		  e depois sobrescrever com as informações que vem no objeto atualizacaoDoTopico */
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if (optional.isPresent()) {
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
 		
-		Topico topico = form.atualizar(id, topicoRepository);
-		return ResponseEntity.ok(new TopicoDto(topico));
+		return ResponseEntity.notFound().build();		
+		/*está vindo um ID do topico que quero atualizar, carregar ele do banco de dados
+		  e depois sobrescrever com as informações que vem no objeto atualizacaoDoTopico */		
+		
 		   /*classe para montar uma resposta*/
 	}
 	
 	@DeleteMapping("/{id}")
 	@Transactional
+	@CacheEvict(value = "listaDeTopicos", allEntries = true) /*limpar o cache*/
 	public ResponseEntity<?> remover(@PathVariable Long id) {
-		topicoRepository.deleteById(id);
-		return ResponseEntity.ok().build();
+		/*tratamento de erro para devolver um erro 404 caso nao exista o id na requisição do cliente*/
+		
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if (optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+		return ResponseEntity.notFound().build();
 	}
 
 }
